pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cerno"
version = "1.0.0"
description = "Modern CLI for Nessus findings review and security tool orchestration"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ridgeback InfoSec, LLC"}
]
keywords = ["nessus", "security", "penetration-testing", "vulnerability-scanning", "nmap", "netexec"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Console",
]

dependencies = [
    "rich>=13.7,<15",
    "typer>=0.12,<1",
    "shellingham>=1.5,<2",
    "colorama>=0.4.6,<1",
    "pyperclip>=1.8,<2",
    "loguru>=0.7.3",
    "requests>=2.32.5",
    "beautifulsoup4>=4.14.2",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "pytest-timeout>=2.1",
    "black>=23.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/ridgebackinfosec/cerno"
Repository = "https://github.com/ridgebackinfosec/cerno"
Issues = "https://github.com/ridgebackinfosec/cerno/issues"

[project.scripts]
cerno = "cerno:app"

[tool.setuptools]
packages = ["cerno_pkg"]
py-modules = ["cerno"]

[tool.setuptools.package-data]
cerno_pkg = ["*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=cerno_pkg",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (database, filesystem)",
    "slow: Slow tests (large file processing)",
]
timeout = 300

[tool.coverage.run]
source = ["cerno_pkg"]
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
